+++ desktop/src/MediaInformationSession.hpp
@@ -103,6 +103,13 @@ public:
   bool isFFprobe() const override;
 
   /**
+   * Returns whether it is an <code>FFplay</code> session or not.
+   *
+   * @return true if it is an <code>FFplay</code> session, false otherwise
+   */
+  bool isFFplay() const override;
+
+  /**
    * Returns whether it is a <code>MediaInformation</code> session or not.
    *
    * @return true if it is a <code>MediaInformation</code> session, false